+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
+## [22.14.4] 2026-05-19
+
+### Fixed
+
+- ML_evaluation: 
+    - Add smart font size adjustment for categorical feature distribution plots in `dit_generation_metrics()` and `autoencoder_metrics()` to prevent overcrowding of x-tick labels when there are many categories.
+    - Improve plot aesthetics for plots in the `autoencoder_metrics()` function, including better spacing, grid lines, and legend formatting.
+
+
 ## [22.14.3] 2026-05-18
 
 ### Changed